requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Cython==3.0.11
dnspython==2.7.0
fastapi==0.115.7
Flask==3.1.0
Flask-Cors==5.0.0
flask-talisman==1.1.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.2
motor==3.6.1
multidict==6.1.0
numpy==1.26.0
packaging==24.2
pandas==2.2.3
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rq==2.1.0
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
